backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cartrap-backend"
version = "0.1.0"
description = "CarTrap backend service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "email-validator>=2.2,<3.0",
  "fastapi>=0.116,<1.0",
  "PyJWT>=2.10,<3.0",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.10,<3.0",
  "pymongo>=4.10,<5.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "mongomock>=4.3,<5.0",
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
